requirements.txt
accelerate
onnxruntime-gpu
numpy
opencv-contrib-python
Pillow
tqdm
matplotlib
transformers
ffmpeg-python
scipy
diffusers[torch]
safetensors
mediapy
